chore: extract charts symbol/grid from dashboard index (move-only)
